+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.0] - 2025-01-29
 ### Added
 - Support for traffic signals detection with a new YoloV11 model.
+- Checks on video and gps durations.
 ### Changed
 - Upgraded ultralytics to 8.3.0.
 - Upgraded documentation.